backend/requirements.txt
# Backend Python dependencies for the MagicScool platform
# Lock specific versions where possible to ensure reproducible builds.
Flask==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
python-dotenv==1.0.0
openai>=1.2.0
pyttsx3>=2.90
pillow>=10.0.0
pytesseract>=0.3.10
pdf2image>=1.17.0
opencv-python-headless>=4.8.0.76
numpy>=1.26.0
werkzeug>=2.3.0